1、将瓦斯泵查询的界面实现分散到EditPumpDBDlg.cpp里，这样MainDlg.cpp内容就少一些，要不然太乱 2、派生实现了一个新的列表控件，为了能够处理双击消息 3、瓦斯泵界面增加了一个删除按钮，原来的界面是处理右键消息，我感觉有点麻烦，加个按钮更省事
